LPS-179214 Extract ScreenNavigationEntry methods from AccountRoleDetailsScreenNavigationCategory to AccountRoleDetailsScreenNavigationEntry, extends AccountRoleDetailsScreenNavigationCategory in it and return getCategoryKey in getEntryKey method
