+++ modules/apps/account/account-admin-web/src/main/java/com/liferay/account/admin/web/internal/frontend/taglib/servlet/taglib/AccountRoleDetailsScreenNavigationCategory.java
@@ -15,25 +15,10 @@
 package com.liferay.account.admin.web.internal.frontend.taglib.servlet.taglib;
 
 import com.liferay.account.admin.web.internal.constants.AccountScreenNavigationEntryConstants;
-import com.liferay.account.admin.web.internal.security.permission.resource.AccountRolePermission;
-import com.liferay.account.constants.AccountRoleConstants;
-import com.liferay.account.model.AccountRole;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
-import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
-import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
 import com.liferay.portal.kernel.language.Language;
-import com.liferay.portal.kernel.model.Role;
-import com.liferay.portal.kernel.model.User;
-import com.liferay.portal.kernel.security.permission.ActionKeys;
-import com.liferay.portal.kernel.security.permission.PermissionCheckerFactoryUtil;
-import com.liferay.portal.kernel.service.RoleLocalService;
-
-import java.io.IOException;
 
 import java.util.Locale;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -42,14 +27,11 @@ import org.osgi.service.component.annotations.Reference;
  * @author Pei-Jung Lan
  */
 @Component(
-	property = {
-		"screen.navigation.category.order:Integer=10",
-		"screen.navigation.entry.order:Integer=10"
-	},
-	service = {ScreenNavigationCategory.class, ScreenNavigationEntry.class}
+	property = "screen.navigation.category.order:Integer=10",
+	service = ScreenNavigationCategory.class
 )
 public class AccountRoleDetailsScreenNavigationCategory
-	implements ScreenNavigationCategory, ScreenNavigationEntry<AccountRole> {
+	implements ScreenNavigationCategory {
 
 	@Override
 	public String getCategoryKey() {
@@ -57,13 +39,8 @@ public class AccountRoleDetailsScreenNavigationCategory
 	}
 
 	@Override
-	public String getEntryKey() {
-		return AccountScreenNavigationEntryConstants.ENTRY_KEY_DETAILS;
-	}
-
-	@Override
 	public String getLabel(Locale locale) {
-		return _language.get(locale, "details");
+		return language.get(locale, "details");
 	}
 
 	@Override
@@ -72,41 +49,7 @@ public class AccountRoleDetailsScreenNavigationCategory
 			SCREEN_NAVIGATION_KEY_ACCOUNT_ROLE;
 	}
 
-	@Override
-	public boolean isVisible(User user, AccountRole accountRole) {
-		if (accountRole == null) {
-			return true;
-		}
-
-		Role role = _roleLocalService.fetchRole(accountRole.getRoleId());
-
-		if ((role != null) && AccountRoleConstants.isSharedRole(role)) {
-			return false;
-		}
-
-		return AccountRolePermission.contains(
-			PermissionCheckerFactoryUtil.create(user),
-			accountRole.getAccountRoleId(), ActionKeys.UPDATE);
-	}
-
-	@Override
-	public void render(
-			HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse)
-		throws IOException {
-
-		jspRenderer.renderJSP(
-			httpServletRequest, httpServletResponse,
-			"/account_entries_admin/account_role/details.jsp");
-	}
-
 	@Reference
-	protected JSPRenderer jspRenderer;
-
-	@Reference
-	private Language _language;
-
-	@Reference
-	private RoleLocalService _roleLocalService;
+	protected Language language;
 
 }
